libs/core/pyproject.toml
[project]
name = "neuraprop-core"
version = "0.1.0"
description = "NeuraProp AI shared core library"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.30.0",
    "alembic>=1.15.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "redis[hiredis]>=5.2.0",
    "boto3>=1.35.0",
    "aioboto3>=13.0.0",
    "structlog>=24.4.0",
    "httpx>=0.28.0",
    "pgvector>=0.3.6",
    "python-jose[cryptography]>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/neuraprop_core"]
